apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.10.6
pydantic-settings==2.1.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# Authentication & OAuth
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
fhir.resources==7.1.0
redis==5.0.1
celery==5.3.4
openpyxl==3.1.2
python-dateutil==2.8.2

# AI/ML Dependencies
scikit-learn==1.5.0
pandas==2.1.4
numpy==1.26.2
prophet==1.1.5

# WhatsApp Notifications
twilio==8.11.1

# Monitoring & Logging
sentry-sdk[fastapi]==2.20.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.27.2
